fix: removing post-build.js file
