+++ packages/apex-lsp-vscode-extension/package.json
@@ -303,10 +303,9 @@
     "precompile": "shx cp -R ../../node_modules/@salesforce/apex-tmlanguage/grammars . && shx mkdir -p out && shx cp package.nls*.json language-configuration.json out/",
     "compile": "tsc --build",
     "bundle": "tsup",
-    "build": "npm run precompile && npm run compile && npm run bundle && npm run post-build",
+    "build": "npm run precompile && npm run compile && npm run bundle",
     "rebuild": "npm run clean && npm run build",
     "copy:files": "shx cp -R ../../node_modules/@salesforce/apex-tmlanguage/grammars .",
-    "post-build": "node scripts/post-build.js",
     "package": "cd dist && vsce package --out ../",
     "test": "jest --passWithNoTests",
     "test:coverage": "jest --coverage --coverageDirectory=./coverage --passWithNoTests",
